requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv
flask
flask-cors
cairosvg 